Rename push protection error dialog file name to match component name
