+++ app/src/models/popup.ts
@@ -23,7 +23,7 @@ import { GitHubRepository } from './github-repository'
 import { ValidNotificationPullRequestReview } from '../lib/valid-notification-pull-request-review'
 import { UnreachableCommitsTab } from '../ui/history/unreachable-commits-dialog'
 import { IAPIComment } from '../lib/api'
-import { ISecretScanResult } from '../ui/secret-scanning/push-protection-error'
+import { ISecretScanResult } from '../ui/secret-scanning/push-protection-error-dialog'
 
 export enum PopupType {
   RenameBranch = 'RenameBranch',
