+++ app/src/ui/secret-scanning/push-protection-error-dialog.tsx
